+++ src/expandableCalendar/weekCalendar.js
@@ -4,12 +4,14 @@ import {FlatList, View, Text} from 'react-native';
 import {Map} from 'immutable';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
+import memoize from 'memoize-one';
 
 import styleConstructor from './style';
 import CalendarList from '../calendar-list';
 import Week from '../expandableCalendar/week';
 import asCalendarConsumer from './asCalendarConsumer';
-import {weekDayNames} from '../dateutils';
+import {weekDayNames, sameWeek} from '../dateutils';
+import {extractComponentProps} from '../component-updater';
 
 const commons = require('./commons');
 const UPDATE_SOURCES = commons.UPDATE_SOURCES;
@@ -55,9 +57,11 @@ class WeekCalendar extends Component {
   }
 
   componentDidUpdate(prevProps) {
-    const {updateSource, date} = this.props.context;
+    const {firstDay, context} = this.props;
+    const {updateSource, date, prevDate} = context;
+    const isSameWeek = sameWeek(date, prevDate, firstDay);
 
-    if (date !== prevProps.context.date && updateSource !== UPDATE_SOURCES.WEEK_SCROLL) {
+    if (date !== prevProps.context.date && updateSource !== UPDATE_SOURCES.WEEK_SCROLL && !isSameWeek) {
       this.setState({items: this.getDatesArray()});
       this.list.current.scrollToIndex({animated: false, index: NUMBER_OF_PAGES});
     }
@@ -91,9 +95,8 @@ class WeekCalendar extends Component {
     return newDate.toString('yyyy-MM-dd');
   }
 
-  getMarkedDates() {
+  getMarkedDates = () => {
     const {context, markedDates} = this.props;
-
     if (markedDates) {
       const marked = _.cloneDeep(markedDates);
 
@@ -105,7 +108,11 @@ class WeekCalendar extends Component {
       return marked;
     }
     return {[context.date]: {selected: true}};
-  }
+  };
+
+  getWeekStyle = memoize((width, style) => {
+    return [{width}, style];
+  });
 
   onDayPress = value => {
     _.invoke(this.props.context, 'setDate', value.dateString, UPDATE_SOURCES.DAY_PRESS);
@@ -177,15 +184,17 @@ class WeekCalendar extends Component {
   };
 
   renderItem = ({item}) => {
-    const {calendarWidth, style, onDayPress, ...others} = this.props;
+    const {style, onDayPress, markedDates, ...others} = extractComponentProps(Week, this.props);
+    const isCurrentWeek = sameWeek(item, this.props.context.date, others.firstDay);
+    const fixedMarkedDates = isCurrentWeek ? this.getMarkedDates() : markedDates;
 
     return (
       <Week
         {...others}
         key={item}
         current={item}
-        style={[{width: calendarWidth || this.containerWidth}, style]}
-        markedDates={this.getMarkedDates()}
+        style={this.getWeekStyle(this.containerWidth, style)}
+        markedDates={fixedMarkedDates}
         onDayPress={onDayPress || this.onDayPress}
       />
     );
@@ -213,10 +222,10 @@ class WeekCalendar extends Component {
     return (
       <View
         testID={this.props.testID}
-        style={[allowShadow && this.style.containerShadow, !hideDayNames && {paddingBottom: 6}]}
+        style={[allowShadow && this.style.containerShadow, !hideDayNames && this.style.containerWrapper]}
       >
         {!hideDayNames && (
-          <View style={[this.style.week, {marginTop: 12, marginBottom: -2}]}>
+          <View style={[this.style.week, this.style.weekCalendar]}>
             {/* {this.props.weekNumbers && <Text allowFontScaling={false} style={this.style.dayHeader}></Text>} */}
             {weekDaysNames.map((day, idx) => (
               <Text